requirements.txt
langchain
faiss-cpu
transformers
sentence-transformers
rich
PyPDF2
python-dotenv
langchain-community
